Changing model to accomodate multiple users view
Before this change I assumed a single-user application. I want to add
support for multiple users, so that we can share projects with each
other.

In this change I'm making ProjectList class specific to a user.
Currently it still loads just person's own project, but that can be
extended to load other public projects.