More clean up.
I'm doing this small improvements in the code base, not because I'm out
of ideas. Quite contrary. I'm thinking intensely about possible backend
solutions, and experimenting in offline mode. While I'm thinking, I don't
want the streak project to be in the same place. My goal is to keep
improving it every day.